Download Pdf added modified location

+++ lib/ui/fragment_customer_details.dart
@@ -3,10 +3,11 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:ui';
 
+import 'package:dio/dio.dart';
+import 'package:ext_storage/ext_storage.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:flutter_file_dialog/flutter_file_dialog.dart';
 import 'package:flutter_grate_app/model/customer_details.dart';
 import 'package:flutter_grate_app/model/estimate.dart';
 import 'package:flutter_grate_app/model/hive/user.dart';
@@ -19,7 +20,7 @@ import 'package:hive/hive.dart';
 import 'package:http/http.dart' as http;
 import 'package:image_picker/image_picker.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
-import 'package:path_provider/path_provider.dart';
+import 'package:permission_handler/permission_handler.dart';
 import 'package:uuid/uuid.dart';
 
 import '../constraints.dart';
@@ -75,7 +76,7 @@ class _CustomerDetailsFragmentState extends State<CustomerDetailsFragment>
 
   List<bool> downloadingList = [];
 
-  var progressString = "";
+  var progressString = "0%";
 
   @override
   void initState() {
@@ -85,22 +86,38 @@ class _CustomerDetailsFragmentState extends State<CustomerDetailsFragment>
     getData();
   }
 
-  Future<void> downloadFile(String url, int index, String invoiceId) async {
-    try{
-    var response = await http.get(url);
-    final output = await getTemporaryDirectory();
-    final file = File("${output.path}/$invoiceId.pdf");
-    await file.writeAsBytes(response.bodyBytes);
-    final params = SaveFileDialogParams(sourceFilePath: "${output.path}/$invoiceId.pdf", localOnly: true);
-    await FlutterFileDialog.saveFile(params: params);
-    setState(() {
-      downloadingList[index] = false;
-      progressString = "Completed";
-    });
-    showDialog(context: context, builder: (context)=>deleteSuccess());
-  } catch(error){
-
-      showDialog(context: context, builder: (context)=>deleteFailed());
+  Future<void> downloadFileNew(String url, int index, String invoiceId) async {
+    var dio = new Dio();
+    if ((await Permission.storage.request().isGranted)) {
+      try {
+        var dir = await ExtStorage.getExternalStorageDirectory();
+        var knockDir =
+            await new Directory('${dir}/GrateAppPdf').create(recursive: true);
+        await dio.download(url, '${knockDir.path}/${invoiceId}.pdf',
+            onReceiveProgress: (rec, total) {
+          //print("Rec: $rec , Total: $total");
+          if (mounted) {
+            setState(() {
+              downloadingList[index] = true;
+              progressString = ((rec / total) * 100).toStringAsFixed(0) + "%";
+            });
+          }
+        });
+        if (mounted) {
+          setState(() {
+            downloadingList[index] = false;
+            print(
+                "File is downloaded to your Internal Storage 'GrateAppPdf' folder!");
+          });
+        }
+        showDialog(context: context, builder: (context) => deleteSuccess());
+        print("Download completed");
+        progressString = "0%";
+      } catch (e) {
+        e.toString();
+      }
+    } else {
+      await Permission.storage.request();
     }
   }
 
@@ -655,7 +672,7 @@ class _CustomerDetailsFragmentState extends State<CustomerDetailsFragment>
                                                     columnWidths: {
                                                       0: FlexColumnWidth(3.25),
                                                       1: FlexColumnWidth(1.75),
-                                                      2: FlexColumnWidth(1.25),
+                                                      2: FlexColumnWidth(1),
                                                     },
                                                     defaultVerticalAlignment:
                                                         TableCellVerticalAlignment
@@ -853,7 +870,7 @@ class _CustomerDetailsFragmentState extends State<CustomerDetailsFragment>
                                                                               index] =
                                                                           true;
                                                                     });
-                                                                    downloadFile(
+                                                                    downloadFileNew(
                                                                         _list[index]
                                                                             .pdfUrl,
                                                                         index,
@@ -880,12 +897,16 @@ class _CustomerDetailsFragmentState extends State<CustomerDetailsFragment>
                                                                               CircleAvatar(
                                                                             backgroundColor:
                                                                                 Colors.grey.shade300,
-                                                                            child:
-                                                                                Icon(
-                                                                              Icons.file_download,
-                                                                              color: Colors.black,
-                                                                              size: 18,
-                                                                            ),
+                                                                            child: downloadingList[index]
+                                                                                ? Text(
+                                                                                    progressString,
+                                                                                    style: Theme.of(context).textTheme.caption.copyWith(color: Colors.black),
+                                                                                  )
+                                                                                : Icon(
+                                                                                    Icons.file_download,
+                                                                                    color: Colors.black,
+                                                                                    size: 18,
+                                                                                  ),
                                                                           ),
                                                                         ),
                                                                         Positioned(
